rename file back to QueryTable
