+++ app/table/QueryTable.tsx
